+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.4
+  * Adjust `client.py` to use API `v.2.0`.
+  * Remove `users` and `leads` streams.
+  * Add `contacts` stream.
+  * Add support for `cursor` style paging.
+
 ## 0.0.3
   * Adjust `client.py` to use API `v.1.4`; fix issues w/ `conversations` and `admins` missinng fields; fix `company_attributes` API version error; fix `write_bookmark` issue when scrolling.